requirements.txt
pandas~=2.2.3
apache-airflow~=3.0.2
psycopg2-binary~=2.9.10
pyodbc~=5.2.0
fastavro~=1.11.1
requests~=2.32.3
numpy~=2.1.3
tenacity~=9.1.2
python-dotenv~=1.1.0
cx_Oracle~=8.3.0
pyspark~=4.0.0
boto3~=1.39.4
botocore~=1.39.4
cx_Oracle~=8.3.0